requirements.txt
streamlit
requests
python-dotenv
bs4
google.generativeai